post delete

+++ app/Http/Controllers/PostsController.php
@@ -102,13 +102,13 @@ class PostsController extends Controller
         return $fileName;
     }
 
-    public function edit(Post $post)
+    public function edit(Request $request, Post $post)
     {
         //$post = Post::find($id); //primary key만 가능
         // $post = Post::where('id', $id)->first();
         // dd($post);
         //수정 폼 생성
-        return view('posts.edit')->with('post', $post);
+        return view('posts.edit', ['post' => $post, 'page' => $request->page]);
     }
     public function update(Request $request, $id) //라우터 파라미터보다 앞에 인덱션 받는 객체가 와야한다
     {
@@ -120,8 +120,19 @@ class PostsController extends Controller
         ]);
         $post = Post::find($id);
         //이미지 파일 수정. 파일 시스템에서
+
+        //Authorization. 즉 권한이 있는지 검사
+        //즉, 로그인한 사용자와 게시글의 작성자가 같은지 체크
+        // if (auth()->user()->id != $post->user_id) {
+        //     abort(403);
+        // }
+
+        if ($request->user()->cannot('update', $post)) {
+            abort(403);
+        }
+
         if ($request->file('imageFile')) {
-            $imagePath = 'public/image/' . $post->image;
+            $imagePath = 'public/images/' . $post->image;
             Storage::delete($imagePath);
             $post->image = $this->uploadPostImage($request);
         }
@@ -130,11 +141,30 @@ class PostsController extends Controller
         $post->content = $request->content;
         $post->save();
 
-        return redirect()->route('post.show', ['id' => $id]);
+        return redirect()->route('post.show', ['id' => $id, 'page' => $request->page]);
     }
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         //파일 시스템에서 이미지 파일 삭제
         //게시글을 데이터베이스에서 삭제
+        $post = Post::findOrFail($id);
+
+        //Authorization. 즉 권한이 있는지 검사
+        //즉, 로그인한 사용자와 게시글의 작성자가 같은지 체크
+        // if (auth()->user()->id != $post->user_id) {
+        //     abort(403);
+        // }
+        if ($request->user()->cannot('delete', $post)) {
+            abort(403);
+        }
+
+        $page = $request->page;
+        if ($post->image) {
+            $imagePath = 'public/images/' . $post->image;
+            Storage::delete($imagePath);
+        }
+        $post->delete();
+
+        return redirect()->route('posts.index', ['page' => $page]);
     }
 }
